+++ tool/generate/builddate/build_date_data.go
@@ -8,11 +8,18 @@ import (
 	"fmt"
 	"github.com/SongZihuan/BackendServerTemplate/tool/generate/basefile"
 	"github.com/SongZihuan/BackendServerTemplate/tool/utils/fileutils"
+	"log"
 	"time"
 )
 
 var buildTime = time.Now()
 
 func WriteBuildDateData() error {
-	return fileutils.Write(basefile.FileBuildDateTxt, fmt.Sprint(buildTime.Unix()))
+	val := fmt.Sprint(buildTime.Unix())
+
+	log.Printf("generate: write build data: %s UTC (timestaamp: %s)\n", buildTime.UTC().Format(time.DateTime), val)
+	defer log.Println("generate: write build data finish")
+
+	log.Printf("generate: write %s to file %s\n", val, basefile.FileBuildDateTxt)
+	return fileutils.Write(basefile.FileBuildDateTxt, val)
 }