+++ tool/generate/releaseinfo/release_data.go
@@ -9,6 +9,7 @@ import (
 	"github.com/SongZihuan/BackendServerTemplate/tool/generate/basefile"
 	"github.com/SongZihuan/BackendServerTemplate/tool/generate/changelog"
 	"github.com/SongZihuan/BackendServerTemplate/tool/generate/git"
+	"log"
 	"os"
 	"text/template"
 )
@@ -32,6 +33,9 @@ func init() {
 }
 
 func WriteReleaseData() error {
+	log.Println("generate: write release info data data")
+	defer log.Println("generate: write write release info data finish")
+
 	file, err := os.OpenFile(basefile.FileReleaseInfoMD, os.O_CREATE|os.O_TRUNC|os.O_WRONLY, 0644)
 	if err != nil {
 		return err
@@ -39,6 +43,20 @@ func WriteReleaseData() error {
 	defer func() {
 		_ = file.Close()
 	}()
+
+	v := git.Version()
+
+	cl := changelog.GetLastChangLog()
+
+	gcl := git.GetGitHubCompareMD()
+
+	log.Printf("generate: release info version: %s\n", v)
+	if gcl != "" {
+		log.Printf("generate: release info github changelog: %s\n", gcl)
+	} else {
+		log.Println("generate: release info github changelog: without")
+	}
+	log.Printf("generate: release info changelog length=%d\n", len(cl))
 
 	return releaseTemplate.Execute(file, templateData{
 		Version:       git.Version(),